:green_heart: update unit tests
