+++ tests/test_archives.py
@@ -136,10 +136,3 @@ class ArchiveTestCases(object):
     def test_implied_dir(self):
         self.fs.getinfo("foo/bar")
         self.fs.getinfo("foo")
-
-    def test_geturl(self):
-        test_file = "foo/bar/egg/foofoo"
-        expected = "zip://{zip_file_path}/{file_inside_zip}".format(
-            zip_file_path=self._temp_path, file_inside_zip=test_file
-        )
-        self.assertEqual(self.fs.geturl(test_file), expected)
